requirements.txt
langchain
langchain-community
qdrant-client
sentence-transformers
FlagEmbedding
torch
unstructured
pypdf
ragas
datasets
tqdm
python-dotenv
pandas
